requirements.txt
asgiref==3.8.1
certifi==2020.12.5
colorama
Django==4.2.17
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna
PyJWT 
requests
sqlparse==0.5.3
urllib3